Ticket #625: More fixes to be able to build FDO against VS2010
git-svn-id: http://svn.osgeo.org/fdo/trunk@5971 d98eaee6-4212-0410-900c-dcc49bf69b99

+++ Utilities/Common/Inc/FdoCommonBinaryReader.h
@@ -100,7 +100,7 @@ class FdoCommonBinaryReader : public FdoDisposable
         void resize()
         {
             size_t old_capacity = m_capacity;
-            m_capacity = (m_capacity == 0) ? 8 : (1.4 * m_capacity);
+            m_capacity = (m_capacity == 0) ? 8 : (size_t)(1.4 * m_capacity);
             CachedBuffer** newArray = new CachedBuffer*[m_capacity];
             for (size_t i = 0; i < old_capacity; i++)
                 newArray[i] = m_buffers[i];
